Dialogue Trigger System Pseudocode Complete
first commit since I started working on it because I forgot to make my own branch this entire time. oops!!

has pseudocode and placeholder variables in place. the trigger filter system for character dialogue is completely set up and just needs to be translated from pseudocode to real code, which is easy but will just take a bit of time.